Fix Installer Choice/Receipt Conflict
The recipe as is will create a pkginfo file that contains an `install_choices_xml` item to exclude `com.google.pkg.Keystone`. However, the receipts array includes this pkg and so the resulting pkginfo file will continually attempt to install.
This change marks the receipt as optional in the resulting pkginfo file.